fix(http): do reverse dns resolution only for IP addresses
We might use an origin fqdn starting with a number.
Thus a reverse dns resolution is requested for a FQDN, which is failing
and adding latency penalty in our test bed (around 20 ms per request).
Refactor some IP checks and use the ats_is_ip helper to determine if the
origin is an IP address.

Signed-off-by: Fatih Acar <facar@scaleway.com>